Cut out the part that generates JSON.
